requirements.txt
numpy==1.16.1
opencv-contrib-python
opencv-python==4.2.0.84
openpyxl==3.1.5
pandas==2.2.3
pillow==11.0.0
pyttsx3



